Revert "restructure for cleaner code"
